refactor:변경된 API 기준으로 인물 조회 API 변경
