+++ src/api/bookApi.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { ENDPOINTS } from "../constants/api";
+import { ENDPOINTS } from "./endpoints";
 
 // API 응답 타입 정의
 export interface BookApiResponse {
@@ -66,17 +66,17 @@ export const getVideosByBookId = async (
   }
 };
 
-// 특정 책의 캐릭터 목록 조회 (GET /books/{bookId}/characters)
-export const getCharactersByBookId = async (
-  bookId: number
-): Promise<CharacterApiResponse[]> => {
-  try {
-    const response = await axios.get<CharacterApiResponse[]>(
-      ENDPOINTS.characters.getByBookId(bookId.toString())
-    );
-    return response.data;
-  } catch (error) {
-    console.error(`Failed to fetch characters for book ${bookId}:`, error);
-    throw error;
-  }
-};
+// // 특정 책의 캐릭터 목록 조회 (GET /books/{bookId}/characters)
+// export const getCharactersByBookId = async (
+//   bookId: number
+// ): Promise<CharacterApiResponse[]> => {
+//   try {
+//     const response = await axios.get<CharacterApiResponse[]>(
+//       ENDPOINTS.characters.getByBookId(bookId.toString())
+//     );
+//     return response.data;
+//   } catch (error) {
+//     console.error(`Failed to fetch characters for book ${bookId}:`, error);
+//     throw error;
+//   }
+// };
